backend/requirements.txt
# FastAPI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0

# Optional dependencies (uncomment if needed)
# psycopg2-binary>=2.9.9  # PostgreSQL (optional)
# alembic>=1.12.0         # Database migrations (optional)
# redis>=5.0.0            # Redis cache (optional)
